PgpBlockViewRenderModule
